nixos/szurubooru: Host option and fixes for backwards compatibility
Pass through alembic and waitress so that it uses the versions from the package
definition and not from the evaluation of the module. This allows the module to
be imported and used even if it is ported to NixOS versions that use a different
default python3 version. This should be more robust against divergence in
version numbers between the system and package.

Also add a `host` config option to set the bind address. Waitress would
otherwise bind to 0.0.0.0, which isn't preferable since it's usually the case
the szurubooru is being run behind a reverse proxy.