nixos/ec2-data.nix: Print whole host pubkeys
This prints the whole host public keys in the same format as cloud-init.
The main benefit is that this allows the direct generation of a known_hosts
file, simplifying client setup.

This commit also provides a new NixOS test that's an alternative to the
currently broken nixos/tests/ec2.nix tests.