stash: Refactor to use finalAttrs, straighten out src references
Previously, it was somewhat difficult to override the source tree for
stash: The `src` attribute was used in several places, some of which 
were unreachable. Now, using the `finalAttrs` argument, both the go
and yarn builds can use the same source tree, overriding it only once.

This change also adds a `postPatch` attribute, whose sole purpose is 
to copy the yarn build result into place, which again makes it easier
to override the source tree without having to duplicate the 
load-bearing other items on `preBuild`.