systemd: use finalAttrs.version
I recently deployed systemd 257.9 to a bunch of machines as mitigation
for #440073. This caused some confusion in the aftermath since we had
257.9 deployed (via an override for `systemd.package` to not rebuild the
world), but `systemctl --version` still reported 257.8.

This patch fixes that problem by using the final version everywhere
instead of the version from the let-binding.

I decided to remove the let-binding entirely to avoid re-introducing the
wrong usage again.