nixos/murmur: Hard-code log directory to /var/log/murmur
The Murmur NixOS module allows logging to a file instead of journald. At
the moment any arbitrary directory may be specified by using the
`logfile` option, but nothing ensures that the directory exists or
proper permissions are configured.

Since there is not much reason to specify a custom directory for log
files anyway, just use `/var/log/murmur` by using the systemd way. This
replaces the option `logfile` with the boolean option `logToFile` as
well.

Signed-off-by: Felix Singer <felixsinger@posteo.net>