linuxKernel.kernels: use lib.makeAlias
This is not a scope, so the callPackage protection doesn't apply. Also,
since these aliases are defined in the `linuxKernel.kernels` attribute
set and not at the top-level, the inherited top-level aliases are not
protected against inclusion in callPackage either. For this to work,
these aliases would just need to be moved to the top-level instead.

However, evaluation with and without config.allowAliases will be the
same here, providing the same safety against actually using these
attributes. Also, of course, they have metadata.

This is mostly to show how a package set could use `lib.mkAlias` - it
will work the same in a proper scope.