ugrep: add optional filter utils wrapping & gnugrep replacement symlinks
The `ugrep+` and `ug+` commands are the same as the
`ugrep` and `ug` commands, but also use filters to
search PDFs, documents, e-books, image metadata,
when these filter tools are present:
- pdftotext
- antiword
- pandoc
- exiftool

`ugrep+` and `ug+` can now be wrapped to make those
utils accessible without the need for users to clutter
their environments through installing them. This is
opt-in with the `wrapWithFilterUtils` flag since these
packages grow the closure size massively.

A `createGrepReplacementLinks` flag was also added to
optionally create symlinks to serve as `gnugrep` drop-in
replacements. This works just like with `coreutils`. We
ensure that these variants beat `gnugrep`'s priority value
so `ugrep` wins when building `system-path` with `buildEnv`.