open5gs{,-webui}: fix webui
As it is now, the built webui won’t work since to be able to start
it, it needs node_modules and the .next directory. Restore the normal
NPM installPhase which will copy all the necessary files.

The webui isn’t used during the main open5gs build, so move it to
its own separate package.