haskellPackages.ghc-typelits-natnormalise: restrict to < 0.8
This seems the best course of action: We are still only performing a
non-breaking update compared to Stackage LTS (which will fix a few
builds) while being able to execute the test suite.

haskellPackages.ghc-tcplugins-extra: 0.4.8 -> 0.5

ghc-typelits-natnormalise requires >= 0.5 since 0.7.11. It does not
look like the 0.5 update actually contains a breaking changes, so
this update is probably safe:

https://github.com/clash-lang/ghc-tcplugins-extra/pull/29#issuecomment-3299008674
https://github.com/clash-lang/ghc-tcplugins-extra/compare/702dda2095c66c4f5148a749c8b7dbcc8a09f5c...v0.5.0