lesscpy: buildPythonApplication -> buildPythonPackage
We are about to add another Python package (jupyter-themes) which
depends on lesscpy, not the lesscpy application, but the lesscpy
Python package. In order to make that package available in the
dependent's Pythons environment, we need to use buildPythonPackage
instead of buildPythonApplication.

See https://github.com/NixOS/nixpkgs/issues/45503 and
https://github.com/NixOS/nixpkgs/commit/b7e0c40d299c for more details.