canokey-qemu: mark broken
Uses a four‐year‐old patched vendored version of Mbed TLS for
cryptography that doesn’t build with CMake 4. Doesn’t build with
current versions of `canokey-core`, either. No upstream development
since 2023.

This was only used in‐tree by the `systemd-initrd-luks-fido2`
NixOS test, which is not a channel blocker and that I couldn’t
find a historical failure for that wasn’t related to issues with
the test driver.