Add a dummy systemd service to the rootless container.
This ensures the network-online.target to be reached, which is normally
not the case without any actual network services. But podman waits for
this target to be reached when starting a rootless container (see
https://github.com/containers/podman/issues/24796 ).

While the tests still work, this change at least removes waiting for a
timeout.

As an aside, it also does not trigger the bug #443423.