treewide(java-users): clean callPackage overrides, by-name some
**This commit should cause no rebuilds.**

# Moving Java overrides out of all-packages

All usages of this pattern have been moved to
'let ... in' blocks.

Patterns I searched for:
* `jdk =`
* `jdk_headless =`
* `jre =`
* `jfx =`

The `jing` alias has been moved to aliases.nix, so
update the usage of `jing` in `xmloscopy` to
`jing-trang`.

cassandra was ignored because that derivation
involves code generation.

I ignored grails, because jdk is set to null by
default by an override in all-packages.nix. That
package would need a slightly larger refactor to
deal with.

Packages with a *-native variant were ignored because
those would need to be refactored.

jetbrains was ignored because that might be a special case?

I ignored virtualbox, because I did not want
to touch the usages of the pattern for other
dependencies.