nixos/virtualisation: add sharedDirectories.<name>.readOnly option
To reduce bandwidth and/or enforce network isolation, a user may wish to
clone and update large repositories on the host. However, since those
repositories may still contain integral information, the user may *also*
desire to disallow the user to write to the repository.

This is exactly what this option solves, assuming that QEMU is used. A
practical application I had right now was for postmarketOS: Cloning [1]
takes rather long on my machine, I don't want to duplicate it for every
builder VM nor them to influence each other.

[1]: https://gitlab.postmarketos.org/postmarketOS/pmaports