linuxManualConfig: Handle only y and m in readConfig
Parsing Linux config files in Nix in full is complicated, primarily due
to string values. Currently, the regular expression used is fragile and
breaks with string values containing equal signs or (escaped) quotes.

The primary use of these options is system.requiredKernelConfig, which
only really needs to track "yes" and "module". Therefore, only parse
those.

This is a breaking change. However, this is a very obscure corner of
linuxManualConfig and should have few if any affected users.