linphonePackages: remove
Linphone is a SIP softphone application, which depends on many bespoke
and custom-patched libraries, collectively referred to as
`linphonePackages` in this commit.

The build process has been failing for a while on both stable and
unstable, but due to the way the packages are organized in nixpkgs,
updating them is quite cumbersome.

Therefore, in an effort to ensure future maintainability, we first
remove all existing packages from nixpkgs, to then be able to later add
them again within their own scope.