haskell.compiler: make krank ignore informational issue references
A lot of the links to issues in pkgs/development/compilers/ghc note
historical issues we need to workaround for outdated versions of GHC
or note related discussion. Them being closed is not actionable in many
cases so I have marked them with krank:ignore-line, so they don't show
up in reports generated using krank.