keycloak: move build step to make derivation reproducible
The build phase was used to do a Quarkus optimisation by caching the
configuration and plugins. Sadly, this re-runs the Hibernate ORM
bytecode generation which uses random strings. It also generates a
properties file with a date in it.

So I've moved the build to the service. It now uses a systemd
StateDirectory to put the generated JAR files and everything else is
linked together into the existing KC_HOME_DIR.

I have tested that Keycloak now starts from scratch, and only rebuilds
when configuration changes.

This fixes both #445665 and #445902