libkdumpfile: init at v0.5.5
Added a fixup patch that is causing a build error. This has been fixed
in the latest master but the latest version (v0.5.5) does not include
this change.

The latest version of this package has moved the python bindings to a
different repo. For now, compile the package without any python
bindings[2].

Once the latest version is released, we can remove both the fixup patch
and configure flags

This package is needed for adding drgn[2] package later.

[1] <https://codeberg.org/ptesarik/pykdumpfile
[2] https://github.com/osandov/drgn