rsync: Remove `fakeroot` from `checkInputs`
Having `fakeroot` in `checkInputs` has the unfortunate side effect of
making rsync depend on systemd, and by extension making the linux
kernel depend on systemd.

```
$ nix why-depends --derivation --all -f . linux systemd.src
/nix/store/ahnjzvyxn3jqaaqmhkrdxfimzd8m6yya-linux-6.12.48.drv
└───/nix/store/07v98114pg0v4vb8xahdj6v488piy23r-rsync-3.4.1.drv
    └───/nix/store/70yyn9xwg3gkmdw7azal22wb8hp2lyxz-fakeroot-1.37.1.2.drv
        └───/nix/store/sac9bwghprnv77v8c65ljrwpr09c2wbi-libcap-2.76.drv
            └───/nix/store/9v3wirxpv9dzjkpr2swkgfwi78q0d72d-linux-pam-1.7.1.drv
                └───/nix/store/gfkzz6c2lhr73habq4x9ci7bj6jfgjgz-systemd-minimal-libs-257.9.drv
                    └───/nix/store/kxpjbfihhzkm2dilaj4xxndxfwmbm98z-source.drv
```

My main goal is keep systemd updates from rebuilding the kernel, and
admittedly the use of rsync in the kernel derivation is completely
unnecessary. But I think, in its own right, removing the dependency of
rsync on systemd is worthwhile.

The consequence of this is that we have to skip two additional tests
(out of the total of 46), namely the `chown` and `devices` tests.