autobase: remove
This has a sodium-native.node with a wrongly set executable bit.
Now, instead of patching it, I'll just drop it because it's not a real
program, but a Node library. We use `buildNpm*` for that purpose.

This only makes maintenance harder because issues like this are
non-trivial to spot!

Related: https://github.com/NixOS/nixpkgs/pull/403638#issuecomment-3339125475