docutils: fix source not fetchable from the UK
fetchFromRepoOrCz silently fails when building packages from the UK.

Attempting to download anything from the mirror whilst your IP is in the UK
will redirect the request to

https://repo.or.cz/uk-blocked.html

which is a notice saying that access from the UK is forbidded due to the
Online Safety Act.

This commit updates the source of docutils so it can be downloaded from PyPI
instead.

See 444342