haskellPackages.mkDerivation: convert any/all cabal files to unix nls
In general, `pname` SHOULD be the same as the Cabal package name in
Haskell updates, but things work fine without this in many cases (e.g.
when only building executables), so we need to preserve this.

Since Cabal only allows a single .cabal file per directory, we can just
process *.cabal.