nixos/shadow: use /bin/sh as default userDefaultShell
The default was previously set by the bash module. It's not set directly
here (to the compiled in value) of the shadow package to allow disabling
the bash module.

Note that this doesn't mean that the effective default shell of a normal
NixOS installation has changed. It is still bash.