linux: get rid of drvAttrs
This is in principle, what Alyssa was already doing[1], but I re-did it
because rebasing this against reformats felt too annoying. The previous
attempt was reverted because of unrelated regressions in the chain.

The `drvAttrs` function made the entire code far more cluttered, harder
to understand and didn't serve a real purpose since it was just used a
single time.

I also got rid of the `kernelConf`-variable and decided to directly use
`stdenv.hostPlatform.linux-kernel`, making it more clear where the
attributes are actually coming from.

[1] f521f4613355f8628fe378dad4c41d3fd8886966

Co-authored-by: Alyssa Ross <hi@alyssa.is>