linux: drop hacky passing around of positions
This is something I introduced in 63185299f4c8beecac7e204dc20394f38e68a477
to make sure the `version`/`src` attributes point to the files that declare
the version such that we get pinged by ofborg for updates.

This is long obsolete since the versions are maintained in a JSON file
now. Also, the kernel team is listed as owner in the ci/OWNERS file for
the entire subtree, so we definitely get pinged by CI nowadays.

Hence, remove the hackery.