fltk14: add libdecor to buildInputs
FLTK_USE_SYSTEM_LIBDECOR is set, but it is not adhered to because libdecor is
not provided in the build environment. This causes FLTK to use bundled libdecor,
which causes an unexpected link-time dependency on GTK+3, which causes a lot of
issues.

Fix this by adding libdecor to the build environment.