gradle: remove dependency on EOL jdk23, use jdk11 for test
A passthru test of the javaToolchains configuration property was 
using jdk23.

There is a requirement that the JDK version for the
test NOT be "the default for any of the gradle versions", so jdk11
is a good choice for the replacement since jdk17 and jdk21 are
currently used and jdk25 will be used in the future.