Reapply "linux: default stdenv.hostPlatform.linux-kernel"
This reverts commit ddaa949afe18949396a1de532df60ffaea5bf2a3.

This was reverted as part of a bulk revert of all kernel changes in a
certain time period — as far as I know there was no actual
justification for reverting this change in particular.  The original
rationale for making this change stands: it's nice for new platforms
not to have to copy the defaults over and over in
lib/systems/platforms.nix.