linux_libre: remove
Closes #447657
Closes #447562
Closes #447563

Ignoring a kernel and whenever it breaks having somebody pop up to do
the update is NOT an acceptable approach to maintaining a kernel
package.

Since there's apparently not enough interest (which is understandable
because that's a kernel that... just removes things), I think a full-blown
removal is the only way forward.

Also, to be clear upfront: the only future this should have is
maintenance in a downstream-project by somebody caring enough. To me, a
re-add is out of question since it'd contradict the "no new downstream
kernel rule".