freecad: temporarily remove broken py-slvs
py-slvs is broken on staging-next due to cmake 4.
It is only required for some plugins in freecad.
The plugin system needs to be reworked, ideally
splitting freecad into a core package and wrapper,
so that python library additions will no longer
rebuild the core package. For now, removing py-slvs
is a bandaid fix.