python3Packages.pyghidra: init at 12.0
The package is available on pypi but it is built from Ghidra
sources[^1]. I packaged it directly from Ghidra sources, so nixpkgs
always has a consistent version of the two.

`pyghidra` can be started both by using the provided cli script or by
importing the packages from a Python program.

[^1]: https://github.com/NationalSecurityAgency/ghidra/blob/master/Ghidra/Features/PyGhidra/src/main/py