monado: Remove protobuf override for opencv
The override was originally added in #442919 because monado depends on
protobuf through two paths:

- monado -> opencv4 -> protobuf
- monado -> onnxruntime -> protobuf

And since, *at that time*, onnxruntime has protobuf pinned to
protobuf_21, we had to override opencv4 with protobuf_21 as well to
avoid mixing two different versions of protobuf. Otherwise, mayhem
occurs on initialization, manifesting as crashes or deadlocks. This does
not seem to be known when #442919 was written.

With #437307, onnxruntime is no longer pinned to protobuf_21, so the
same mixing problem was introduced back into monado, just the other
direction.

Revert the override for opencv4 to fix it.