release.nix: Use arm64-apple-darwin in bootstrapConfigs, not aarch64
Fix this config name for consistency with non-bootstrap platform config.
This also fixes a build failure of the bootstrap tools on
aarch64-darwin, where a test in LLVM 21 assumes it's spelt arm64, not
aarch64.

Also saves on having to rebuild LLVM just for bootstrap tools, now that
the config is consistent.