cosmic-greeter: add missing cosmic-randr and orca dependency
The path to the orca screen reader was hardcoded to /usr/bin/orca, which prevents it from being executed in a NixOS environment.

This change adds orca as a dependency and patches the source code to use the Nix-provided executable path, allowing the greeter to launch it correctly.

Also adds cosmic-randr to PATH.