nixos/users-groups: move home directory creation to systemd-tmpfiles
The creation of user home directories was previously handled by an
activation script. This could lead to a race condition where home
directories were created in the /home mountpoint directory before the
/home filesystem was mounted.

This commit refactors this logic to use `systemd-tmpfiles` instead.
A `tmpfiles.d` rule is now generated for each user with `createHome = true`.
`systemd-tmpfiles` runs after filesystems are mounted, which resolves
the race condition.

This approach is based on the original work by John Soo in #223932

Co-authored-by: John Soo <john.soo@arista.com>