gradle: use official Gradle wrapper
Not using the official Gradle wrapper causes papercuts like GRADLE_OPTS
and JAVA_HOME not working or misbehaving. Use the default wrapper script
and allow override of JAVA_HOME so we preserve the default Gradle wrapper
behavior.

Move the Gradle jars and bin to $out/libexec while we're at it so we
don't have to make any modifications to the wrapper ourselves
except for patching the shebang on install.

Tested with `gradle_{7,8,9}.tests` and a `nixpkgs-review rev HEAD`
to check reverse dependencies.

Fixes: #72220
Fixes: #402297
Fixes: #453296