gradle: use official Gradle wrapper
Not using the default Gradle wrapper causes papercuts like GRADLE_OPTS
not working or misbehaving: see #72220. Use the default wrapper script
so we preserve the default Gradle wrapper behavior.

Note that this shouldn't be breaking unless users depended on one of
the default Gradle options not being set by the Nix Gradle wrapper.