dockerTools: Refactor common writing code into writeImageStream
Previously, the buildLayeredImage and buildNixShellImage functions
duplicated image compression functionality that was used to store the
output of streamLayeredImage and streamNixShellImage.

This patch factors out the common code into a combined function:
writeImageStream.

Doing this also makes it far more practical to customize and override
the behaviour of streamLayeredImage and streamNixShellImage, because
buildLayeredImage and buildNixShellImage formed an opaque wrapping
around the stream functions.

Therefore, this patch deprecates buildLayeredImage and
buildNixShellImage, and encourages users to write images by combining
one of the stream functions with writeImageStream.

Co-authored-by: commiterate <111539270+commiterate@users.noreply.github.com>