nexusmods-app: switch back to `fetchFromGitHub`
We switched to `fetchgit` because we needed to fetch git-LFS contend for
some tests to build. Since then upstream has stopped using git-LFS and
we have stopped fetching with it enabled, so we can switch back to
`fetchFromGitHub`.

I assume that when we disabled git-LFS, I wasn't aware `fetchFromGitHub`
supported `fetchSubmodules`.