nixos/timesyncd: remove unnecessary workaround
This workardoung is now unnecessary because systemd and
systemd-timesyncd will never actually revert to a time lower than the
current system time. It will only *advance* the system time to the EPOCH
if the system time is lower.

This also makes systemd-timesyncd bashless as we remove the preStart
script for any NixOS version newer than 19.09.