m1n1: replace `withBranding` with `customLogo` parameter
This avoids having to use `overrideAttrs` to customize the logo
installation and instead places the interface as an argument which
represents a file path to the desired image which will be
resized by imagemagick to 128x128 and 256x256.

This means that the same image will be used for both sizes (which
should be the intended case) but that can be circumvented manually via
`overrideAttrs` if desired (rather than "when needed").

It also adds documentation for this argument within
`meta.longDescription` for m1n1.