treewide: Comment out `webkitgtk_4_0`
`webkitgtk_4_0` was already transitively insecure because of `libsoup_2_4` and it will be removed soon.

Ideally, we would just replace it with `libsoup_3`-based `webkitgtk_4_1`
but it is unclear if they would work with the 4.1 ABI. At best, the programs
would only be looking for 4.0, failing to build or start, at worst there
might be conflicts with libsoup 2 symbols if that is pulled in by anything
or crashes that will only happen after certain action.

Let’s mark the packages as broken to ensure human intervention.