treewide: Replace `webkitgtk_4_0` with `webkitgtk_4_1`
`webkitgtk_4_0` was already transitively insecure because of `libsoup_2_4` and it will be removed soon.
Let’s switch to `libsoup_3`-based `webkitgtk_4_1`.

Also mark the packages as broken since it is unclear if they will work with the 4.1 ABI.
At best, the programs will be looking for 4.0, failing to build or run,
at worst there might be conflicts with libsoup 2 symbols if that is pulled in by anything.