linuxPackages.virtualboxGuestAdditions: Add depmod config
The 3rdparty VirtualBox guest kernel modules conflict with the ones in
the mainline Linux kernel, which is why the mainline modules are not
enabled in the NixOS kernel configuration. In order to resolve the
conflict, add a configuration for depmod allowing to override the
search paths for these modules. So, if this package is added to
`boot.extraModulePackages` then depmod will read the configuration and
it will use these modules instead of the mainline ones.

Signed-off-by: Felix Singer <felixsinger@posteo.net>