workflows/{review,reviewed}: rename from dismissed-review and review-dismissed
This is just a separate commit for a better diff. This workflow will be
extended with the `submitted` type in the next commit, so needs a
different naming.

While at it, also remove the useless concurrency group: This workflow is
never run in a PR context, so will never be cancelled.

(cherry picked from commit f1640b71d81c06ce30378758c0ff5bb977105756)