nixos/release-notes: add notes about Nextcloud
Use the same template as every year to guide people through the upgrade.

Also, put the stuff into the NixOS release notes: the only deployment
way for the package that we truly support is via the module. I think
this shows again what a remarkably bad idea this separation is IMHO.